requirements.txt
openai-whisper
faster-whisper==0.10.1
SpeechRecognition
soundfile
gradio

--extra-index-url https://download.pytorch.org/whl/cu118

torch
torchvision
torchaudio
